requirements.txt
Django==1.10
django-autocomplete-light>=2.0,<3.0
django-materialize-forms==0.2.0
num2words==0.5.3
numpy==1.11.1
pandas==0.18.1
Pillow==3.3.1
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
django-storages==1.5.1
boto==2.43.0
requests==2.11.1
